pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "allure-step-rewriter"
version = "0.2.0"
description = "Rewrite Allure step titles without nesting"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Nikita Tulenkov", email = "tulenckov.nikita@gmail.com"}
]
keywords = ["allure", "pytest", "testing", "automation", "qa", "step", "rewrite"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Testing",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: Pytest",
    "Operating System :: OS Independent",
]

dependencies = []

[project.optional-dependencies]
# For users who want to install allure-pytest automatically
allure = [
    "allure-pytest>=2.9.0",
]

# For development
dev = [
    "allure-pytest>=2.9.0",
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

# Install everything at once
all = [
    "allure-pytest>=2.9.0",
]

[project.urls]
Homepage = "https://github.com/NikitaTule/allure-step-rewriter"
Documentation = "https://github.com/NikitaTule/allure-step-rewriter#readme"
Repository = "https://github.com/NikitaTule/allure-step-rewriter"
"Bug Tracker" = "https://github.com/NikitaTule/allure-step-rewriter/issues"

[tool.setuptools]
packages = ["allure_step_rewriter"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=allure_step_rewriter --cov-report=html --cov-report=term"

[tool.coverage.run]
source = ["allure_step_rewriter"]
omit = ["tests/*", "examples/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false